refactor(exo): split exo tools
