+++ packages/exo/index.js
@@ -1,8 +1,6 @@
 export {
-  defendPrototype,
-  defendPrototypeKit,
   initEmpty,
   defineExoClass,
   defineExoClassKit,
   makeExo,
-} from './src/interface-tools.js';
+} from './src/exo-makers.js';
