+++ packages/exo/package.json
@@ -18,6 +18,7 @@
   "module": "./index.js",
   "exports": {
     ".": "./index.js",
+    "./tools.js": "./tools.js",
     "./package.json": "./package.json"
   },
   "scripts": {
