+++ packages/exo/tools.js
@@ -1,0 +1,1 @@
+export { defendPrototype, defendPrototypeKit } from './src/exo-tools.js';
